kitm/requirements.txt
APScheduler==3.11.0
Django==5.1.4
django-apscheduler==0.7.0
django-bootstrap5==24.3
django-cleanup==9.0.0
django-debug-toolbar==4.4.6
flake8==7.1.1
psycopg2-binary==2.9.10
pytest==7.4.4
pytest-cov==6.0.0
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-subtests==0.9.0
python-dotenv==1.0.1
Pillow
requests
